+++ CHANGELOG.md
@@ -12,6 +12,13 @@ All notable changes to this project will be documented in this file.
     .then((tasks) => console.table(tasks))
   ```
 - Added `SqlJsDataProvider` for use with front end sqlite implementation [sql.js](https://sql.js.org/)
+- Added support for return value for validations - true || undefined are valid, string will provide the message. For example:
+  ```ts
+  @Fields.string({
+    validate:(task)=>task.title>5 || "too short"
+  })
+  ```
+-
 - Fixed issue where `findOne` didn't work
 
 ## [0.24.1] 2023-12-30